+++ git-goals/plugins/examples/hello_world.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+# Example plugin: Hello World
+
+git_goals_hello_world() {
+    echo "Hello from the git-goals plugin system!"
+}
+
+# Register the new subcommand
+git_goals_register_command "hello" "git_goals_hello_world" "Print a hello message"